diagnostics: split metrics/mod.rs into transaction/histogram/error files
mod.rs was 1300+ lines. Split per the established domain groupings:

- transaction_metrics.rs absorbs LoadKind, QueryType, TransactionOutcome,
  TransactionLifecycleCounters, TransactionLifecycleSnapshot — all
  transaction-shape concepts the existing observer types already reference.
- histogram_metrics.rs holds the HistogramMetrics primitive +
  HistogramUnit + HistogramSnapshot. DatabaseHistograms (in mod.rs)
  composes them.
- error_metrics.rs holds ErrorMetrics + ErrorInfo.

ServerProperties / ServerMetrics stay in mod.rs (too small to split).
LoadMetrics, DatabaseMetrics, *LoadMetrics, DatabaseHistograms,
ClientEndpoint stay in mod.rs. ActionMetrics stays in mod.rs.
